Give root variable names a prefix
Before the change there many of the variables had names which easily
conflicted with what a user of the library might have in their code
already. To prevent overriding the values of the user (or vice versa)
this change updates the variables with the 'ms' (short for
MultiSelect) prefix to prevent common names from conflicting.